+++ client/index.html
@@ -1,14 +1,16 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Notifikasi</title>
-</head>
-<body>
-    <h1><center>Selamat datang di Terbang Tinggi App</center></h1>
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Terbang Tinggi App</title>
+  </head>
+  <body>
+    <h1><center>Selamat datang di API Terbang Tinggi App</center></h1>
+    <h3><center>Backend : Achmad, John, Umar</center></h3>
+    <h3><center>Frontend : Aroyan, Frans, Rizky</center></h3>
 
     <script src="./client.js"></script>
-</body>
+  </body>
 </html>
