+++ controllers/index.js
@@ -1,4 +1,5 @@
 const user = require("./user");
+const auth = require("./auth");
 const webpush = require("./webpush");
 
-module.exports = { user, webpush };
+module.exports = { user, auth, webpush };
