+++ routes/index.js
@@ -1,7 +1,9 @@
 const express = require("express");
 const router = express.Router();
 const user = require("./user");
+const auth = require("./auth");
 
-router.use("/auth", user);
+router.use("/auth", auth);
+router.use("/user", user);
 
 module.exports = router;
